#85983 - move configure eda command to a new namespace
